Refactoring for UI layout for different spoke stack options
